move flyto logic to Popup.ts, refs #7
