+++ src/lib/components/navbar/content/News.svelte
@@ -1,39 +1,15 @@
 <script lang="ts" context="module">
-	import geojson from '../../../../../data/app-geojson.json';
-	import { map } from '$lib/store';
+	import { geodata } from '$lib/store';
 	import { get } from 'svelte/store';
-	import { createPopup, type FeatureForPopup } from '$lib/components/Popup';
 	import { instanceOfFeatureForPopup } from '$lib/components/Popup';
+	import { flyTo } from '$lib/components/Popup';
 
-	// check if geojson.features is Feature[] using instanceOfFeatureForPopup
-	// note that geojson.features is array of feature, so we need to map over it
+	// check if geodata.features is Feature[] using instanceOfFeatureForPopup
+	// note that geodata.features is array of feature, so we need to map over it
 	// and check each feature
-	const newFeatures = geojson.features
-		.slice(0, 5) // latest 5 features
-		.filter((feature) => instanceOfFeatureForPopup(feature)) as FeatureForPopup[];
-
-	function flyTo(feature: FeatureForPopup) {
-		const popup = createPopup(feature.geometry.coordinates, feature.properties);
-
-		const [lng, lat] = feature.geometry.coordinates;
-
-		const _map = get(map);
-
-		_map.flyTo({
-			speed: 1,
-			curve: 1,
-			easing(t) {
-				return t;
-			},
-			essential: true,
-			center: [lng, lat],
-			zoom: 15
-		});
-		// wait for flyTo
-		setTimeout(() => {
-			popup.addTo(_map);
-		}, 1000);
-	}
+	const newFeatures = get(geodata)
+		.features.slice(0, 5) // latest 5 features
+		.filter((feature) => instanceOfFeatureForPopup(feature));
 </script>
 
 <ul>
